[ty] Move `KnownInstanceType`, and related types, to a new `known_instance.rs` submodule
