+++ crates/ty_python_semantic/src/types/class.rs
@@ -33,6 +33,7 @@ use crate::types::generics::{
     GenericContext, InferableTypeVars, Specialization, walk_specialization,
 };
 use crate::types::infer::{infer_expression_type, infer_unpack_types, nearest_enclosing_class};
+use crate::types::known_instance::DeprecatedInstance;
 use crate::types::member::{Member, class_member};
 use crate::types::mro::DynamicMroError;
 use crate::types::relation::{HasRelationToVisitor, IsDisjointVisitor, TypeRelation};
@@ -43,9 +44,9 @@ use crate::types::visitor::{TypeCollector, TypeVisitor, walk_type_with_recursion
 use crate::types::{
     ApplyTypeMappingVisitor, Binding, BindingContext, BoundSuperType, CallableType,
     CallableTypeKind, CallableTypes, DATACLASS_FLAGS, DataclassFlags, DataclassParams,
-    DeprecatedInstance, FindLegacyTypeVarsVisitor, IntersectionBuilder, KnownInstanceType,
-    MaterializationKind, PropertyInstanceType, TypeContext, TypeMapping, TypedDictParams,
-    UnionBuilder, VarianceInferable, binding_type, declaration_type, determine_upper_bound,
+    FindLegacyTypeVarsVisitor, IntersectionBuilder, KnownInstanceType, MaterializationKind,
+    PropertyInstanceType, TypeContext, TypeMapping, TypedDictParams, UnionBuilder,
+    VarianceInferable, binding_type, declaration_type, determine_upper_bound,
 };
 use crate::{
     Db, FxIndexMap, FxIndexSet, FxOrderSet, Program,
